refactor the installer and entitlement
