+++ FileAuditSystem/FileAuditSystem/AppDelegate.swift
@@ -10,17 +10,18 @@ import Cocoa
 @main
 class AppDelegate: NSObject, NSApplicationDelegate, InstallerDelegate {
    
-
     @IBOutlet var window: NSWindow!
     
     let folderManageVC = FolderManageViewController()
     let checkSextVC = CheckExtensionViewController()
-
+    
     func isExtensionUpAndRunning() -> Bool {
+        // TODO: in future, check if alreay running
+        // as of now always return false, as not running, let user manually install every time app runs
         return false;
     }
     
-    func applicationDidFinishLaunching(_ aNotification: Notification) {
+    func applicationDidFinishLaunching(_ aNotification: Notification) {        
         if isExtensionUpAndRunning() {
             self.window.contentView?.addSubview(folderManageVC.view)
         }
@@ -35,15 +36,5 @@ class AppDelegate: NSObject, NSApplicationDelegate, InstallerDelegate {
         checkSextVC.view.removeFromSuperview()
         self.window.contentView?.addSubview(folderManageVC.view)
     }
-    
-    func applicationWillTerminate(_ aNotification: Notification) {
-        // Insert code here to tear down your application
-    }
-    
-    func applicationSupportsSecureRestorableState(_ app: NSApplication) -> Bool {
-        return true
-    }
-
-
 }
